Support multiple descriptors.
